Renamed constants module, removed import * statement
