+++ .gitignore
@@ -5,3 +5,4 @@ build/
 model.tflite
 model.tflite.tmp
 .idea/
+venv/
