+++ ulozto_downloader/const.py
