+++ ulozto_downloader/utils.py
@@ -1,6 +1,6 @@
 import sys
 import colors
-from .constants import *
+from . import const
 
 
 def _print(text, x=0, y=0):
@@ -11,8 +11,8 @@ def _print(text, x=0, y=0):
 
 
 def print_part_status(id, text):
-    _print(colors.blue(f"[Part {id}]") + f"\t{text}", y=(id + CLI_STATUS_STARTLINE))
+    _print(colors.blue(f"[Part {id}]") + f"\t{text}", y=(id + const.CLI_STATUS_STARTLINE))
 
 
 def print_captcha_status(text, parts):
-    _print(colors.blue("[CAPTCHA solve]") + f"\t{text}", y=(parts + 2 + CLI_STATUS_STARTLINE))
+    _print(colors.blue("[CAPTCHA solve]") + f"\t{text}", y=(parts + 2 + const.CLI_STATUS_STARTLINE))
